cmake: Use target export feature
This eliminates the need for a "Findcppunit.cmake" script in projects
needing the cppunit library. The project configuration files are
installed along with the library itself and provide information about
where to find the library and how to use it.
A prominent example of this method are the Qt5 libraries, which are
shipping CMake project configuration files by default.

Change-Id: I9db4f07d4a56361215eb68d9080955fede8eb205
Signed-off-by: Bernhard Sessler <bernhard.sessler@corscience.de>
Reviewed-on: https://gerrit.libreoffice.org/7112
Reviewed-by: Markus Mohrhard <markus.mohrhard@googlemail.com>
Tested-by: Markus Mohrhard <markus.mohrhard@googlemail.com>